pyproject.toml
[tool.poetry]
name = "onchain-mas-hedge-fund"
version = "0.1.0"
description = "Onchain MAS Hedge Fund"
authors = ["choguun <visaruth.s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.14"
langchain-openai = "^0.3.0"
langgraph = "^0.2.62"
pandas = "^2.2.3"
numpy = "^2.2.1"
python-dotenv = "^1.0.1"
matplotlib = "^3.10.0"
tabulate = "^0.9.0"
colorama = "^0.4.6"
hyperliquid-python-sdk = "^0.9.0"
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
